refactor: enhance SplittedLayout with TokenSkeleton
